customer_service_agent/pyproject.toml
[project]
name = "customer-service-agent"
version = "0.1.0"
description = "Customer Service Agent component"
requires-python = ">=3.10"
dependencies = [
    "google-adk[eval]",
    "pandas",
    "google-cloud-aiplatform[agent_engines,evaluation]>=1.69.0",
    "google-adk",
    "python-dotenv",
]
